Fix git stats
